server/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
requests==2.32.3
urllib3<2
python-dotenv==1.0.1
yt-dlp>=2024.10.22
SQLAlchemy>=2.0.40,<3.0.0
psycopg2-binary>=2.9.10,<3.0.0
httpx>=0.27.0,<1.0.0
